move out the fifo cache impl
